syscalls/cma: fix the failure of compiling with O2 option
This issue is triggered by the commit 9b02cd465f70 which I wrote. If
compile the process_vm_readv03.c with O2 option and run, it will return
TFAIL. By casting the int type to unsigned long type in the ltp_syscall
calling functions can we avoid this failure. Thus I modify all the
testcases which might be affected by the previous commit.

[jstancek: 9b02cd465f70 removed a function that handled int->long
promotion. glibc's syscall.S takes 6th argument from stack
as quadword.
  ENTRY (syscall)
          movq %rdi, %rax         /* Syscall number -> rax.  */
          movq %rsi, %rdi         /* shift arg1 - arg5.  */
          movq %rdx, %rsi
          movq %rcx, %rdx
          movq %r8, %r10
          movq %r9, %r8
          movq 8(%rsp),%r9        /* arg6 is on the stack.  */
          syscall                 /* Do the system call.  */
If only 'int' is pushed by syscall() then the result depends on
other stack content. If it happens to be != 0, then "flags" is
non-zero and we get EINVAL from kernel.]

+++ testcases/kernel/syscalls/cma/process_vm_readv03.c
@@ -181,7 +181,7 @@ static long *fetch_remote_addrs(void)
 	remote.iov_len = len;
 
 	TEST(ltp_syscall(__NR_process_vm_readv, pids[0], &local,
-			 1, &remote, 1, 0));
+			 1UL, &remote, 1UL, 0UL));
 	if (TEST_RETURN != len)
 		tst_brkm(TFAIL | TERRNO, tst_exit, "process_vm_readv");
 
@@ -213,8 +213,9 @@ static void child_invoke(int *bufsz_arr)
 	tst_resm(TINFO, "child 1: %d local iovecs initialized.",
 		 NUM_LOCAL_VECS);
 
-	TEST(ltp_syscall(__NR_process_vm_readv, pids[0], local, NUM_LOCAL_VECS,
-				   remote, nr_iovecs, 0));
+	TEST(ltp_syscall(__NR_process_vm_readv, pids[0], local,
+			    (unsigned long)NUM_LOCAL_VECS, remote,
+			    (unsigned long)nr_iovecs, 0UL));
 	if (TEST_RETURN != bufsz)
 		tst_brkm(TBROK | TERRNO, tst_exit, "process_vm_readv");
 
@@ -248,7 +249,7 @@ static void setup(void)
 	tst_require_root();
 
 	/* Just a sanity check of the existence of syscall */
-	ltp_syscall(__NR_process_vm_readv, getpid(), NULL, 0, NULL, 0, 0);
+	ltp_syscall(__NR_process_vm_readv, getpid(), NULL, 0UL, NULL, 0UL, 0UL);
 
 	nr_iovecs = nflag ? SAFE_STRTOL(NULL, nr_opt, 1, IOV_MAX) : 10;
 	bufsz = sflag ? SAFE_STRTOL(NULL, sz_opt, NUM_LOCAL_VECS, LONG_MAX)
